move currency selector
