+++ src/components/BankDetails/useCurrencies.ts
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import { useGetCurrenciesMutation } from "services/aws/bankDetails";
 import { useErrorContext } from "contexts/ErrorContext";
 import { EMAIL_SUPPORT } from "constants/env";
-import { Currency } from "./CurrencySelector";
+import { Currency } from "./WiseCurrencies/CurrencySelector";
 
 const DEFAULT_TARGET_CURRENCY = "USD";
